refactor: extract SessionCoreRepository
